Überarbeitung des Package model.coinvertion
-X, Y, Z und Spin Speed Convertion klasse erzeugt, sind aber noch nicht
fertig
-AngleToSpeed klasse muss auch noch verarbeitet werden
-LeapMotion Klasse erzeugt und momentan soweit fertig

+++ src/com/quadromotion/gestures/IGestures.java
@@ -2,16 +2,16 @@ package com.quadromotion.gestures;
 
 public interface IGestures {
 
-	public float getPitchRightHand();
-	public float getRollRightHand();
-	public float getYawRightHand();
-	public float getThrustRightHand();
-	public float getSphereRadiusRightHand();
+	public double getPitchRightHand();
+	public double getRollRightHand();
+	public double getYawRightHand();
+	public double getThrustRightHand();
+	public double getSphereRadiusRightHand();
 	
-	public float getPitchLeftHand();
-	public float getRollLeftHand();
-	public float getYawLeftHand();
-	public float getThrustLeftHand();
-	public float getSpehreRadiusLeftHand();
+	public double getPitchLeftHand();
+	public double getRollLeftHand();
+	public double getYawLeftHand();
+	public double getThrustLeftHand();
+	public double getSpehreRadiusLeftHand();
 
 }
